fix OptionsService fix libtheora x86 add MENGINE_PLATFORM_WINDOWS64 add MENGINE_PLATFORM_WINDOWS32 improve devplugins

+++ src/Systems/DX9RenderSystem/DX9ErrorHelper.cpp
@@ -82,7 +82,7 @@ namespace Mengine
             , m_line
             , m_method
             , message
-            , _hr
+            , (uint32_t)_hr
             );
 
         return true;
